fixed Issues based on feedback from mobitar
-The "Enable TouchID" button is grayed out and looks unselectable. I
think you should use the same style as the other buttons. That is, it
should always be red text, but the text should be either
enable/disable.(DONE)

-When the app is locked, it should not show notes list, it should show
account tab instead. (DONE)

-Even with TouchID enabled, I was able to select "Export Data" in
between the alerts. This is a major issue, since anyone can export my
data even when I've enabled TouchID. (DONE)

-For code style, I think we should try to copy existing style as much
as possible, to make it more uniform. This means a lot less space in
between lines of code. (DONE)