Add broken transition to composer
The ViewControllers are not usable on their own and need to be displayed
by an embedded navigation controller. When instantiating a view controller
without pushing it through a navigation controller, the topBar is missing,
leaving the user without a [Done] button.

A first step solution that doesnt require structural changes to the storyboard
would be to gain access to the current displaying viewController and make it push
the composer viewcontroller.
However, an optional binding check for NotesViewController on the selectedViewController-Property
of the apps tabbarcontroller fails.